refactor: major refactor restructuring project
